Refactor QueuePin, mobile responsiveness, fix click out
+ QueuePin component
* Wrap pin when on mobile
* Move ref to div element above input to fix checkmark functionality
